requirements.txt
ollama
python-dotenv
PyMuPDF
